chore: updated endpoints with auth guard

+++ src/home/home.controller.ts
@@ -8,9 +8,18 @@ import {
   Post,
   Put,
   Query,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { HomeService } from './home.service';
-import { CreateHomeDto, HomeResponseDto, UpdateHomeDto } from './dto/home.dto';
+import {
+  CreateHomeDto,
+  HomeResponseDto,
+  InquireDto,
+  UpdateHomeDto,
+} from './dto/home.dto';
+import { User, UserType } from 'src/user/decorators/user.decorator';
+import { Roles } from 'src/decorators/roles.decorators';
+import { UserType as UserT } from '@prisma/client';
 
 @Controller('home')
 export class HomeController {
@@ -43,21 +52,56 @@ export class HomeController {
     return this.homeService.getHome(id);
   }
 
+  @Roles(UserT.ADMIN, UserT.REALTOR)
   @Post()
-  createHome(@Body() body: CreateHomeDto) {
-    return this.homeService.createHome(body);
+  createHome(@Body() body: CreateHomeDto, @User() user: UserType) {
+    return this.homeService.createHome(body, user?.id);
   }
 
+  @Roles(UserT.ADMIN, UserT.REALTOR)
   @Put(':id')
-  updateHome(
+  async updateHome(
     @Param('id', ParseIntPipe) id: number,
     @Body() body: UpdateHomeDto,
+    @User() user: UserType,
   ) {
+    const realtor = await this.homeService.getRealtorByHomeId(user?.id);
+
+    if (realtor.realtor.id !== user?.id) throw new UnauthorizedException();
     return this.homeService.updateHome(body, id);
   }
 
+  @Roles(UserT.ADMIN, UserT.REALTOR)
   @Delete(':id')
-  deleteHome(@Param('id', ParseIntPipe) id: number) {
+  async deleteHome(
+    @Param('id', ParseIntPipe) id: number,
+    @User() user: UserType,
+  ) {
+    const realtor = await this.homeService.getRealtorByHomeId(user?.id);
+
+    if (realtor.realtor.id !== user?.id) throw new UnauthorizedException();
     return this.homeService.deleteHome(id);
+  }
+
+  @Roles(UserT.ADMIN, UserT.REALTOR, UserT.BUYER)
+  @Post('/inquire/:id')
+  inquire(
+    @Param('id', ParseIntPipe) id: number,
+    @User() user: UserType,
+    @Body() { message }: InquireDto,
+  ) {
+    return this.homeService.inquire(user, id, message);
+  }
+
+  @Roles(UserT.ADMIN, UserT.REALTOR, UserT.BUYER)
+  @Get('/:id/messages')
+  async getHomeMessages(
+    @Param('id', ParseIntPipe) id: number,
+    @User() user: UserType,
+  ) {
+    const realtor = await this.homeService.getRealtorByHomeId(user?.id);
+
+    if (realtor.realtor.id !== user?.id) throw new UnauthorizedException();
+    return this.homeService.getHomeMessages(id);
   }
 }
